Added gridsearch notebook, pdf
Gridserach here is actually a randomized search, but the idea is the same. The
PDF has the saved output of a 30-fold cv with five different models. Results
are interesting. This shows me, though, that CV can't handled threshholded
regressors because it doesn't automatically do the `y_hat = y_hat - yhat.min()`
and `y_hat = y_hat / y_hat.max()` to rescale it from 0 to 1.