Added preprocessing script and updated notebook
The preprocessing script is a work in progress. It currently does not work
because it performs different feature transformations on train and test
respectively because the script is looking at qualities of the set and
generatiung features based on that data. So if train and test are different in,
say, the number of violation codes which appear fewer than 10 times, each set
will have a different set of features, making them incompatible.